requirements.txt
# =============================================================================
# Investment Analyst API - Main Requirements
# =============================================================================

# Core FastAPI and async support
fastapi>=0.128.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# Core Python utilities
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
typing-extensions>=4.8.0
numpy
pandas>=2.0.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.20.0

# LangChain Core & Extensions
langchain>=0.3.0
langchain-core>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-qdrant>=0.2.0

# LangGraph & Checkpointing
langgraph>=1.0.0
langgraph-checkpoint>=3.0.0
langgraph-checkpoint-sqlite>=3.0.0
langgraph-prebuilt>=1.0.0
langgraph-sdk>=0.2.0

# Stock Analysis Multi-Agent System (Quant)
langgraph-supervisor>=0.0.31
langchain-mcp-adapters>=0.2.1

# MCP (Model Context Protocol) Dependencies
fastmcp>=0.4.0
mcp>=1.9.2

# Vector Database & Search
qdrant-client>=1.7.0
fastembed>=0.3.0

# LLM Providers & Embeddings
openai>=1.0.0
langchain-groq
langchain_tavily
tavily-python>=0.3.0

# Stock Market Data & Analysis
yfinance>=0.2.40
pandas-ta  # Alternative technical analysis library
ta>=0.11.0  # Technical Analysis library

# Natural Language Processing
textblob>=0.18.0
nltk  # Required by textblob

# Data Visualization
plotly>=5.18.0
kaleido>=0.2.1

# Scientific Computing & Statistics
scipy>=1.11.0
scikit-learn  # For advanced analytics

# Google AI (for chart analysis)
google-generativeai>=0.3.0

# Parsing & Document Processing
llama_parse

# Media & Cloud Storage
cloudinary

# HTTP & Async Utilities
requests>=2.31.0
httpx>=0.26.0
aiohttp>=3.9.0

# Testing (optional)
pytest>=7.4.3


# =============================================================================
# Integration System Dependencies
# =============================================================================
# These are for data source connectors (S3, SharePoint, Google Drive, etc.)
# Can also install separately with: pip install -r requirements_integrations.txt

# Google Drive Integration
google-api-python-client>=2.116.0
google-auth>=2.27.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-generativeai
# SharePoint / Microsoft Graph Integration
msal>=1.26.0  # Microsoft Authentication Library

# Azure Blob Storage Integration
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# AWS S3 Integration
boto3>=1.34.0
botocore>=1.34.0

# SFTP Integration
paramiko>=3.4.0
cryptography>=42.0.0

# =============================================================================
# Notes:
# =============================================================================
# - For MCP servers (Stock Analysis), see quant/stock_agent/README.md
# - MCP servers run separately on ports 8565, 8566, 8567
# - Main API runs on port 8000
# - All dependencies are compatible and tested togetherpip